[BPK-3973] Migrate slider examples to new format
